@@ -4,37 +4,32 @@
   "version": "1.1.3",
   "author": "Austin Green",
   "dependencies": {
-    "bulma": "^0.8.0",
-    "disqus-react": "^1.0.8",
-    "gatsby": "^5.3.3",
-    "gatsby-image": "^2.3.5",
-    "gatsby-plugin-feed": "^2.5.11",
-    "gatsby-plugin-google-tagmanager": "^2.3.5",
-    "gatsby-plugin-netlify": "^2.2.4",
-    "gatsby-plugin-netlify-cms": "^7.3.0",
-    "gatsby-plugin-purgecss": "^6.1.2",
-    "gatsby-plugin-react-helmet": "^3.2.5",
-    "gatsby-plugin-sass": "^6.3.1",
-    "gatsby-plugin-sharp": "^5.3.2",
-    "gatsby-remark-copy-linked-files": "^2.2.4",
-    "gatsby-remark-images": "^3.2.6",
-    "gatsby-remark-prismjs": "^3.5.6",
-    "gatsby-remark-relative-images": "^0.3.0",
-    "gatsby-source-filesystem": "^5.3.1",
-    "gatsby-transformer-remark": "^6.3.2",
-    "gatsby-transformer-sharp": "^5.3.1",
-    "lodash": "^4.17.21",
+    "bulma": "^0.9.0",
+    "gatsby": "^4.0.0",
+    "gatsby-plugin-image": "^2.0.0",
+    "gatsby-plugin-netlify": "^5.0.0",
+    "gatsby-plugin-netlify-cms": "^6.0.0",
+    "gatsby-plugin-purgecss": "^6.0.0",
+    "gatsby-plugin-react-helmet": "^5.0.0",
+    "gatsby-plugin-sass": "^5.0.0",
+    "gatsby-plugin-sharp": "^4.0.0",
+    "gatsby-remark-copy-linked-files": "^5.0.0",
+    "gatsby-remark-images": "^6.0.0",
+    "gatsby-remark-relative-images": "^2.0.2",
+    "gatsby-source-filesystem": "^4.0.0",
+    "gatsby-transformer-remark": "^5.0.0",
+    "gatsby-transformer-sharp": "^4.0.0",
+    "lodash": "^4.17.15",
     "lodash-webpack-plugin": "^0.11.4",
-    "netlify-cms-app": "^2.9.6",
-    "netlify-cms-media-library-cloudinary": "^1.3.2",
-    "netlify-cms-media-library-uploadcare": "^0.5.2",
-    "node-sass": "^7.0.3",
-    "prismjs": "^1.27.0",
+    "netlify-cms-app": "^2.15.72",
+    "netlify-cms-media-library-cloudinary": "^1.3.10",
+    "netlify-cms-media-library-uploadcare": "^0.5.10",
     "prop-types": "^15.6.0",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-helmet": "^6.0.0",
-    "uuid": "^7.0.0"
+    "sass": "^1.43.2",
+    "uuid": "^8.0.0"
   },
   "keywords": [
     "gatsby"
@@ -46,10 +41,14 @@
     "start": "npm run develop",
     "build": "npm run clean && gatsby build",
     "develop": "npm run clean && gatsby develop",
+    "serve": "gatsby serve",
     "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"{gatsby-*.js,src/**/*.js}\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
     "prettier": "^2.0.5"
+  },
+  "engines": {
+    "node": ">= 14.15.0"
   }
 }
