@@ -1,105 +1,79 @@
-import React from 'react'
-import { Link } from 'gatsby'
-import github from '../img/github-icon.svg'
-import logo from '../img/ilyanaDev.svg'
-import twitter from '../img/social/twitter.svg'
+import React, { useState } from "react";
+import { Link } from "gatsby";
+import github from "../img/github-icon.svg";
+import logo from "../img/logo.svg";
 
-const Navbar = class extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      active: false,
-      navBarActiveClass: '',
-    }
-  }
+const Navbar = () => {
+  const [isActive, setIsActive] = useState(false);
 
-  toggleHamburger = () => {
-    // toggle the active boolean in the state
-    this.setState(
-      {
-        active: !this.state.active,
-      },
-      // after state has been updated,
-      () => {
-        // set the class in state for the navbar accordingly
-        this.state.active
-          ? this.setState({
-              navBarActiveClass: 'is-active',
-            })
-          : this.setState({
-              navBarActiveClass: '',
-            })
-      }
-    )
-  }
-
-  render() {
-    return (
-      <nav
-        className="navbar is-transparent"
-        role="navigation"
-        aria-label="main-navigation"
-      >
-        <div className="container">
-          <div className="navbar-brand">
-            <Link to="/" className="navbar-item" title="Logo">
-              <img src={logo} alt="ilyanaDev" style={{ width: '150px' }} />
-            </Link>
-            {/* Hamburger menu */}
-            <div
-              className={`navbar-burger burger ${this.state.navBarActiveClass}`}
-              data-target="navMenu"
-              onClick={() => this.toggleHamburger()}
-            >
-              <span />
-              <span />
-              <span />
-            </div>
-          </div>
-          <div
-            id="navMenu"
-            className={`navbar-menu ${this.state.navBarActiveClass}`}
+  return (
+    <nav
+      className="navbar is-transparent"
+      role="navigation"
+      aria-label="main-navigation"
+    >
+      <div className="container">
+        <div className="navbar-brand">
+          <Link to="/" className="navbar-item" title="Logo">
+            <img src={logo} alt="Kaldi" style={{ width: "88px" }} />
+          </Link>
+          {/* Hamburger menu */}
+          <button
+            className={`navbar-burger burger ${isActive && "is-active"}`}
+            aria-expanded={isActive}
+            onClick={() => setIsActive(!isActive)}
           >
-            <div className="navbar-start has-text-centered">
+            <span />
+            <span />
+            <span />
+          </button>
+        </div>
+        <ul id="navMenu" className={` navbar-start has-text-centered navbar-menu ${isActive && "is-active"}`}>
+            {/* TODO: inline override of padding is a result of refactoring
+                to a ul for accessibilty purposes, would like to see a css
+                re-write that makes this unneccesary.
+             */}
+            <li className="navbar-item" style={{padding: "0px"}}>
               <Link className="navbar-item" to="/about">
                 About
               </Link>
-              
-              <Link className="navbar-item" to="/blog">
-                Blog
-              </Link>
-              <Link className="navbar-item" to="/contact">
-                Contact
-              </Link>
-              
-            </div>
-            <div className="navbar-end has-text-centered">
+            </li>
+            <li className="navbar-item" style={{padding: "0px"}}>
+            <Link className="navbar-item" to="/products">
+              Products
+            </Link>
+            </li>
+            <li className="navbar-item" style={{padding: "0px"}}>
+            <Link className="navbar-item" to="/blog">
+              Blog
+            </Link>
+            </li>
+            <li className="navbar-item" style={{padding: "0px"}}>
+            <Link className="navbar-item" to="/contact">
+              Contact
+            </Link>
+            </li>
+            <li className="navbar-item" style={{padding: "0px"}}>
+            <Link className="navbar-item" to="/contact/examples">
+              Form Examples
+            </Link>
+            </li>
+          <li className="navbar-end has-text-centered">
             <a
-                className="navbar-item"
-                href="https://twitter.com/ilyanaDev"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <span className="icon">
-                  <img src={twitter} alt="Twitter" />
-                </span>
-              </a>
-              <a
-                className="navbar-item"
-                href="https://github.com/ilyanaDev?tab=repositories"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <span className="icon">
-                  <img src={github} alt="Github" />
-                </span>
-              </a>
-            </div>
-          </div>
-        </div>
-      </nav>
-    )
-  }
-}
+              className="navbar-item"
+              href="https://github.com/netlify-templates/gatsby-starter-netlify-cms"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <span className="icon">
+                <img src={github} alt="Github" />
+              </span>
+            </a>
+          </li>
+        </ul>
+      </div>
+    </nav>
+  );
+};
 
-export default Navbar
+export default Navbar;
