@@ -1,26 +1,35 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import Img from 'gatsby-image'
+import * as React from "react";
+import PropTypes from "prop-types";
+import { GatsbyImage } from "gatsby-plugin-image";
 
 const PreviewCompatibleImage = ({ imageInfo }) => {
-  const imageStyle = { borderRadius: '5px' }
-  const { alt = '', childImageSharp, image } = imageInfo
+  const imageStyle = { borderRadius: "5px" };
+
+  const { alt = "", childImageSharp, image } = imageInfo;
 
   if (!!image && !!image.childImageSharp) {
     return (
-      <Img style={imageStyle} fluid={image.childImageSharp.fluid} alt={alt} />
-    )
+      <GatsbyImage
+        image={image.childImageSharp.gatsbyImageData}
+        style={imageStyle}
+        alt={alt}
+      />
+    );
+  } else if (!!childImageSharp) {
+    return (
+      <GatsbyImage
+        image={childImageSharp.gatsbyImageData}
+        style={imageStyle}
+        alt={alt}
+      />
+    );
+    // for Netlify CMS 
+  } else if (image) {
+    return <img style={{imageStyle}} src={image} alt={alt} />;
+  } else {
+    return null
   }
-
-  if (!!childImageSharp) {
-    return <Img style={imageStyle} fluid={childImageSharp.fluid} alt={alt} />
-  }
-
-  if (!!image && typeof image === 'string')
-    return <img style={imageStyle} src={image} alt={alt} />
-
-  return null
-}
+};
 
 PreviewCompatibleImage.propTypes = {
   imageInfo: PropTypes.shape({
@@ -29,6 +38,6 @@ PreviewCompatibleImage.propTypes = {
     image: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired,
     style: PropTypes.object,
   }).isRequired,
-}
+};
 
-export default PreviewCompatibleImage
+export default PreviewCompatibleImage;
