@@ -1,6 +1,6 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { v4 } from 'uuid'
+import * as React from "react";
+import PropTypes from "prop-types";
+import { v4 } from "uuid";
 
 const Testimonials = ({ testimonials }) => (
   <div>
@@ -14,7 +14,7 @@ const Testimonials = ({ testimonials }) => (
       </article>
     ))}
   </div>
-)
+);
 
 Testimonials.propTypes = {
   testimonials: PropTypes.arrayOf(
@@ -23,6 +23,6 @@ Testimonials.propTypes = {
       author: PropTypes.string,
     })
   ),
-}
+};
 
-export default Testimonials
+export default Testimonials;
