@@ -1,24 +1,26 @@
-import React from 'react'
-import { Helmet } from 'react-helmet'
-import { Link, graphql } from 'gatsby'
-import Layout from '../components/Layout'
+import * as React from "react";
+import { Helmet } from "react-helmet";
+import { Link, graphql } from "gatsby";
+import Layout from "../components/Layout";
 
-class TagRoute extends React.Component {
-  render() {
-    const posts = this.props.data.allMarkdownRemark.edges
+const TagRoute = (props) =>  {
+
+    const posts = props.data.allMarkdownRemark.edges;
+
     const postLinks = posts.map((post) => (
       <li key={post.node.fields.slug}>
         <Link to={post.node.fields.slug}>
           <h2 className="is-size-2">{post.node.frontmatter.title}</h2>
         </Link>
       </li>
-    ))
-    const tag = this.props.pageContext.tag
-    const title = this.props.data.site.siteMetadata.title
-    const totalCount = this.props.data.allMarkdownRemark.totalCount
+    ));
+
+    const { tag } = props.pageContext;
+    const { title } = props.data.site.siteMetadata;
+    const { totalCount } = props.data.allMarkdownRemark;
     const tagHeader = `${totalCount} post${
-      totalCount === 1 ? '' : 's'
-    } tagged with “${tag}”`
+      totalCount === 1 ? "" : "s"
+    } tagged with “${tag}”`;
 
     return (
       <Layout>
@@ -28,7 +30,7 @@ class TagRoute extends React.Component {
             <div className="columns">
               <div
                 className="column is-10 is-offset-1"
-                style={{ marginBottom: '6rem' }}
+                style={{ marginBottom: "6rem" }}
               >
                 <h3 className="title is-size-4 is-bold-light">{tagHeader}</h3>
                 <ul className="taglist">{postLinks}</ul>
@@ -40,11 +42,10 @@ class TagRoute extends React.Component {
           </div>
         </section>
       </Layout>
-    )
-  }
+    );
 }
 
-export default TagRoute
+export default TagRoute;
 
 export const tagPageQuery = graphql`
   query TagPage($tag: String) {
@@ -71,4 +72,4 @@ export const tagPageQuery = graphql`
       }
     }
   }
-`
+`;
